braid/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "braid"
version = "1.0.0"
authors = [
    {name = "BRAID Development Team"},
]
description = "BRAID: Block Records Arrow Indexed Dataset - A Python library for reading chunked segmentation data from sharded Arrow files"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["braid", "segmentation", "arrow", "chunks", "neuroglancer", "compression", "neuroscience"]

dependencies = [
    "numpy>=1.20.0",
    "pyarrow>=10.0.0",
    "zstandard>=0.18.0",
]

[project.optional-dependencies]
gcs = ["google-cloud-storage>=2.0.0"]
dev = [
    "pytest>=6.0",
    "pytest-cov>=3.0",
    "black>=22.0",
    "isort>=5.0",
    "flake8>=4.0",
    "mypy>=0.900",
]

[project.urls]
Homepage = "https://github.com/your-org/braid"
Repository = "https://github.com/your-org/braid.git"
Issues = "https://github.com/your-org/braid/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=braid --cov-report=term-missing"